pyproject.toml
[project]
name = "personal-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-store-mongodb>=0.0.1",
    "nest-asyncio>=1.6.0",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "trustcall>=0.0.39",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
# As we are using flat layout we need to explicitly list packages
packages = ["app"]

[tool.mypy]
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
exclude = []
explicit_package_bases = true
warn_no_return = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
check_untyped_defs = false # keep this setting only for tests folder
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["app"]
testpaths = ["tests"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = true
indent-style = "space"

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = ["notebooks"]
src = ["app"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
split-on-trailing-comma = false
